you should not use a comma here
Comma Use
If you are using a conjunction to connect only two items, it is incorrect to use a comma before the conjunction. In addition, if you are using a conjunction to add a phrase that cannot stand alone as a complete sentence, it is incorrect to use a comma before the conjunction.



Instead of: Meng, and Kim are hiking across Ireland.
Consider: Meng and Kim are hiking across Ireland.


Instead of: Two books of fiction, and a book of poetry were on the table.
Consider: Two books of fiction and a book of poetry were on the table.